+++ src/op/mod.rs
@@ -122,6 +122,7 @@ pub mod is_regular;
 pub mod is_simple;
 pub mod is_subdigraph;
 pub mod is_symmetric;
+pub mod is_walk;
 pub mod iter_arcs;
 pub mod iter_out_neighbors;
 pub mod iter_out_neighbors_mut;
@@ -149,6 +150,7 @@ pub use {
     is_simple::IsSimple,
     is_subdigraph::IsSubdigraph,
     is_symmetric::IsSymmetric,
+    is_walk::IsWalk,
     iter_arcs::IterArcs,
     iter_out_neighbors::IterOutNeighbors,
     iter_out_neighbors_mut::IterOutNeighborsMut,